Deduplicate ia sync endpoints into link.py, share client IP helper
- Merge /api/unlink into link.py: one module now hosts both ia sync
  endpoints, the shared HMAC auth dependency (verify_ia_sync_msg /
  IaSyncMsgDep), and the ImportBot save helper.
- Extract get_client_ip / ClientIpDep into fastapi/utils.py and adopt
  it in link.py, importapi.py, and merge_authors.py, replacing three
  inline X-Forwarded-For snippets.
- merge_authors now records only the first X-Forwarded-For hop as the
  edit IP (previously the whole header).
- Unlink 401 responses now return {"detail": "Unauthorized"} like
  link, instead of an empty body.
