+++ openlibrary/fastapi/importapi.py
@@ -9,8 +9,9 @@ from typing import Annotated
 from fastapi import APIRouter, Form, Query
 
 from openlibrary.fastapi.auth import LibrarianDep  # noqa: TC001
+from openlibrary.fastapi.utils import ClientIpDep  # noqa: TC001
 from openlibrary.plugins.importapi.import_ui import ImportPreviewRequest
-from openlibrary.utils.request_context import req_context, web_ctx_ip
+from openlibrary.utils.request_context import web_ctx_ip
 
 router = APIRouter(tags=["import"])
 
@@ -41,6 +42,7 @@ def _build_preview_response(
 @router.get("/import/preview.json")
 def import_preview_json_get(
     _: LibrarianDep,
+    client_ip: ClientIpDep,
     source: Annotated[
         str | None,
         Query(
@@ -65,14 +67,13 @@ def import_preview_json_get(
 
     Requires admin, librarian, or super-librarian role.
     """
-    x_fwd = req_context.get().x_forwarded_for
-    client_ip = x_fwd.split(",")[0].strip() if x_fwd else "127.0.0.1"
     return _build_preview_response(source, provider, identifier, save=False, client_ip=client_ip)
 
 
 @router.post("/import/preview.json")
 def import_preview_json_post(
     _: LibrarianDep,
+    client_ip: ClientIpDep,
     source: Annotated[str | None, Form()] = None,
     provider: Annotated[str | None, Form()] = None,
     identifier: Annotated[str | None, Form()] = None,
@@ -83,6 +84,4 @@ def import_preview_json_post(
 
     Requires admin, librarian, or super-librarian role.
     """
-    x_fwd = req_context.get().x_forwarded_for
-    client_ip = x_fwd.split(",")[0].strip() if x_fwd else "127.0.0.1"
     return _build_preview_response(source, provider, identifier, save=save, client_ip=client_ip)
